refactor(quality): decompose debt_scorecard_validation.py (952→119 LOC)
Split the monolithic validation module into focused sub-modules:
- _primitives.py (92 LOC): shared parsers and type validators
- _baseline_validation.py (148 LOC): baseline + registry groups
- _governance_validation.py (84 LOC): governance section
- _quarterly_targets_validation.py (120 LOC): quarterly targets
- _grace_windows_validation.py (174 LOC): grace windows + allowances
- _decomposition_validation.py (309 LOC): owner/expiry decomposition,
  priority burndown, program done criteria

The orchestrator re-exports public API for backward compatibility.
Removed the now-unnecessary 980 LOC exemption from architecture metrics.

RF-003: Decompose large files > 700 LOC.
